requirements.txt
python-dotenv
fastapi
uvicorn
pydantic[email]